requirements.txt
aiohttp==3.7.4.post0
aiomysql==0.0.21
aiopg==1.3.1
aiosqlite==0.17.0
async-timeout==3.0.1
attrs==21.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.4
chatbot==1.5.2b0
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
click==7.1.2
clock==0.1
cog==0.0.3
Cogs==0.4.4
colorama==0.4.4
cryptography==2.8
DateTime==4.3
dbus-python==1.2.16
discord-components==2.1.1
discord.py==1.7.3
discordSuperUtils==0.1.9
distro==1.4.0
docopt==0.6.2
entrypoints==0.3
fasteners==0.14.1
ffmpeg==1.4
future==0.18.2
fuzzywuzzy==0.18.0
h11==0.9.0
h2==3.2.0
hentai==3.2.8
hpack==3.0.0
hstspreload==2021.9.1
httplib2==0.14.0
hyperframe==5.2.0
idna==2.10
joblib==1.0.1
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
mathparse==0.1.2
monotonic==1.5
motor==2.5.1
multidict==5.1.0
mutagen==1.45.1
netifaces==0.10.4
nltk==3.6.2
oauthlib==3.1.0
olefile==0.46
packaging==21.0
paramiko==2.6.0
peewee==3.13.3
pexpect==4.6.0
Pillow==8.3.1
Pint==0.17
pipreqs==0.4.10
protobuf==3.6.1
prsaw==0.4.0
psycopg2-binary==2.9.1
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
PyGObject==3.42.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==3.12.0
PyMySQL==0.9.3
PyNaCl==1.4.0
pyparsing==2.4.7
pyRFC3339==1.1
python-dateutil==2.7.5
pytz==2021.1
pyxdg==0.26
PyYAML==3.13
randomstuff.py==2.0.0
regex==2021.8.28
reportlab==3.5.34
requests==2.26.0
requests-unixsocket==0.2.0
rfc3986==1.5.0
SecretStorage==2.3.1
sentry-sdk==0.19.4
simplejson==3.16.0
six==1.16.0
sniffio==1.2.0
spotify-dl==7.6.0
spotipy==2.16.1
SQLAlchemy==1.2.19
systemd-python==234
tqdm==4.62.2
typer==0.3.2
typing-extensions==3.10.0.0
tzdata==2021.1
tzlocal==3.0
urllib3==1.26.6
wadllib==1.3.3
yarg==0.1.9
yarl==1.6.3
youtube-dl==2021.6.6
zope.interface==5.4.0
